isize

+++ core/src/allocate.rs
@@ -1,6 +1,6 @@
 use crate::Dinero;
 
-pub fn allocate(item: &Dinero, ratios: Vec<i64>) -> Vec<Dinero> {
+pub fn allocate(item: &Dinero, ratios: Vec<isize>) -> Vec<Dinero> {
     // TODO this is a placeholder implementation
     ratios
         .iter()
